+++ src/utils/events.ts
@@ -1,7 +1,7 @@
 import EventEmitter from 'eventemitter3';
 
 /**
- * Note: we could potentially use a context provider for notifications, but this is fairly lightweight and makes it easy to send errors to sentry.
+ * Note: we could potentially use a context provider for notifications, but this is fairly lightweight.
  */
 const eventEmitter = new EventEmitter();